auth service

+++ auth_service/src/proto/auth.rs
@@ -5,8 +5,8 @@ pub struct Token {
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct UserId {
-    #[prost(int64, tag = "1")]
-    pub user_id: i64,
+    #[prost(int32, tag = "1")]
+    pub user_id: i32,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Nothing {
